+++ app/build.gradle
@@ -3,9 +3,8 @@ apply plugin: 'com.android.application'
 android {
     compileSdkVersion 21
     buildToolsVersion "21.1.2"
-
     defaultConfig {
-        applicationId "uk.ac.lancs.aurorawatch.aurorawatchapp"
+        applicationId 'org.aurorawatchdevs.aurorawatchapp'
         minSdkVersion 9
         targetSdkVersion 21
         versionCode 1
@@ -16,6 +15,8 @@ android {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
+    }
+    productFlavors {
     }
 }
