+++ app/src/main/java/org/aurorawatchdevs/aurorawatch/AlertLevel.java
@@ -1,0 +1,11 @@
+package org.aurorawatchdevs.aurorawatch;
+
+/**
+ * Created by jamesb on 11/02/2015.
+ */
+public enum AlertLevel {
+    none,
+    minor,
+    amber,
+    red
+}
