--- app/src/main/java/uk/ac/lancs/aurorawatch/AlertLevel.java
@@ -1,11 +1,0 @@
-package uk.ac.lancs.aurorawatch;
-
-/**
- * Created by jamesb on 11/02/2015.
- */
-public enum AlertLevel {
-    none,
-    minor,
-    amber,
-    red
-}
