+++ app/src/main/res/layout/activity_main.xml
@@ -2,7 +2,7 @@
 <android.support.v4.widget.DrawerLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools" android:id="@+id/drawer_layout"
     android:layout_width="match_parent" android:layout_height="match_parent"
-    tools:context="uk.ac.lancs.aurorawatch.activity.MainActivity">
+    tools:context="org.aurorawatchdevs.aurorawatch.activity.MainActivity">
 
     <!-- As the main content view, the view below consumes the entire
          space available using match_parent in both dimensions. -->
@@ -19,7 +19,7 @@
     <fragment android:id="@+id/navigation_drawer"
         android:layout_width="@dimen/navigation_drawer_width" android:layout_height="match_parent"
         android:layout_gravity="start"
-        android:name="uk.ac.lancs.aurorawatch.fragment.NavigationDrawerFragment"
+        android:name="org.aurorawatchdevs.aurorawatch.fragment.NavigationDrawerFragment"
         tools:layout="@layout/fragment_navigation_drawer" />
 
 </android.support.v4.widget.DrawerLayout>
