+++ app/src/main/res/layout/fragment_navigation_drawer.xml
@@ -3,4 +3,4 @@
     android:layout_height="match_parent" android:choiceMode="singleChoice"
     android:divider="@android:color/transparent" android:dividerHeight="0dp"
     android:background="#cccc"
-    tools:context="uk.ac.lancs.aurorawatch.fragment.NavigationDrawerFragment" />
+    tools:context="org.aurorawatchdevs.aurorawatch.fragment.NavigationDrawerFragment" />
