+++ app/src/main/res/menu/main.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="uk.ac.lancs.aurorawatch.activity.MainActivity">
+    tools:context="org.aurorawatchdevs.aurorawatch.activity.MainActivity">
 
     <item android:id="@+id/action_settings"
         android:icon="@drawable/ic_action_settings"
