+++ app/src/main/res/values/strings.xml
@@ -24,7 +24,7 @@
     <string name="formatErrorTitle">Unrecognised format</string>
     <string name="formatError">Sorry, the latest information from Aurora Watch UK is not in a recognised format. If this error persists, please update the app to the latest version.</string>
     <string name="formatErrorCheckForUpdates">Check for updates</string>
-    <string name="marketURL">market://details?id=uk.ac.lancs.aurorawatch</string>
+    <string name="marketURL">market://details?id=org.aurorawatchdevs.aurorawatch</string>
     <string name="awUrlButton">AuroraWatch UK Website</string>
     <string name="faq">FAQs</string>
     <string name="awTwitter">AuroraWatch UK on Twitter</string>
